Shift tasks related methods from main class to task class
Methods related to Tasks should not be kept in main.

Move tasks related methods from Main.java to Task.java.

This better follows the convention of OOP.
