+++ src/main/java/Task.java
@@ -1,3 +1,5 @@
+import java.util.ArrayList;
+
 public class Task {
     private String name;
     private boolean done;
@@ -21,5 +23,21 @@ public class Task {
 
     public void setDone(boolean done) {
         this.done = done;
+    }
+
+    public static void addToTasksList(String input, ArrayList<Task> listOfTasks) {
+        System.out.println("added: "+input);
+        Task task = new Task(input, false);
+        listOfTasks.add(task);
+    }
+
+    public static void listTasks(ArrayList<Task> listOfTasks) {
+        for (int i = 0; i < listOfTasks.size(); i++) {
+            if (listOfTasks.get(i).isDone()) {
+                System.out.println((i + 1) + ".[X] " + listOfTasks.get(i).getName());
+            } else {
+                System.out.println((i + 1) + ".[ ] " + listOfTasks.get(i).getName());
+            }
+        }
     }
 }
